Move `*SimulationResultsHandle` to respective inner-classes
