+++ merlin-server/src/main/java/gov/nasa/jpl/aerie/merlin/server/remotes/postgres/PostgresResultsCellRepository.java
@@ -13,8 +13,8 @@ import gov.nasa.jpl.aerie.merlin.protocol.types.ValueSchema;
 import gov.nasa.jpl.aerie.merlin.server.ResultsProtocol;
 import gov.nasa.jpl.aerie.merlin.server.ResultsProtocol.State;
 import gov.nasa.jpl.aerie.merlin.server.models.PlanId;
-import gov.nasa.jpl.aerie.merlin.server.models.PostgresSimulationResultsHandle;
 import gov.nasa.jpl.aerie.merlin.server.models.ProfileSet;
+import gov.nasa.jpl.aerie.merlin.server.models.SimulationResultsHandle;
 import gov.nasa.jpl.aerie.merlin.server.models.Timestamp;
 import gov.nasa.jpl.aerie.merlin.server.remotes.ResultsCellRepository;
 import org.apache.commons.lang3.tuple.Pair;
@@ -25,6 +25,7 @@ import org.slf4j.LoggerFactory;
 import javax.sql.DataSource;
 import java.sql.Connection;
 import java.sql.SQLException;
+import java.time.Instant;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
@@ -221,7 +222,7 @@ public final class PostgresResultsCellRepository implements ResultsCellRepositor
     }
   }
 
-  public static List<Triple<Integer, String, ValueSchema>> getSimulationTopics(Connection connection, long datasetId)
+  private static List<Triple<Integer, String, ValueSchema>> getSimulationTopics(Connection connection, long datasetId)
   throws SQLException
   {
     try (final var getSimulationTopicsAction = new GetSimulationTopicsAction(connection)) {
@@ -229,7 +230,7 @@ public final class PostgresResultsCellRepository implements ResultsCellRepositor
     }
   }
 
-  public static SortedMap<Duration, List<EventGraph<Pair<Integer, SerializedValue>>>>
+  private static SortedMap<Duration, List<EventGraph<Pair<Integer, SerializedValue>>>>
   getSimulationEvents(
       final Connection connection,
       final long datasetId,
@@ -241,7 +242,7 @@ public final class PostgresResultsCellRepository implements ResultsCellRepositor
     }
   }
 
-  public static Pair<Map<SimulatedActivityId, SimulatedActivity>, Map<SimulatedActivityId, UnfinishedActivity>> getActivities(
+  private static Pair<Map<SimulatedActivityId, SimulatedActivity>, Map<SimulatedActivityId, UnfinishedActivity>> getActivities(
       final Connection connection,
       final long datasetId,
       final Timestamp startTime
@@ -485,6 +486,79 @@ public final class PostgresResultsCellRepository implements ResultsCellRepositor
         // A dataset should only be deleted by its cell
         throw new Error("Cell references nonexistent simulation dataset");
       }
+    }
+  }
+
+  public static class PostgresSimulationResultsHandle implements SimulationResultsHandle {
+
+    SimulationDatasetRecord record;
+    DataSource dataSource;
+
+    public PostgresSimulationResultsHandle(DataSource dataSource, SimulationDatasetRecord record) {
+      this.dataSource = dataSource;
+      this.record = record;
+    }
+
+    public SimulationResults getSimulationResults() {
+      try (final var connection = this.dataSource.getConnection()) {
+        final var startTimestamp = record.simulationStartTime();
+        final var simulationStart = startTimestamp.toInstant();
+        final var simulationDuration = Duration.of(
+            startTimestamp.microsUntil(record.simulationEndTime()),
+            Duration.MICROSECONDS);
+
+        final var profiles = ProfileRepository.getProfiles(connection, record.datasetId());
+        final var activities = getActivities(connection, record.datasetId(), startTimestamp);
+        final var topics = getSimulationTopics(connection, record.datasetId());
+        final var events = getSimulationEvents(connection, record.datasetId(), startTimestamp);
+
+        return new SimulationResults(
+            ProfileSet.unwrapOptional(profiles.realProfiles()),
+            ProfileSet.unwrapOptional(profiles.discreteProfiles()),
+            activities.getLeft(),
+            activities.getRight(),
+            simulationStart,
+            simulationDuration,
+            topics,
+            events
+        );
+      } catch (SQLException e) {
+        throw new RuntimeException(e);
+      }
+    }
+
+    @Override
+    public ProfileSet getProfiles(final Iterable<String> profileNames) {
+      try (final var connection = this.dataSource.getConnection()) {
+        return ProfileRepository.getProfiles(connection, record.datasetId(), profileNames);
+      } catch (SQLException e) {
+        throw new RuntimeException(e);
+      }
+    }
+
+    @Override
+    public Map<SimulatedActivityId, SimulatedActivity> getSimulatedActivities() {
+      try (final var connection = this.dataSource.getConnection()) {
+        final var activities = getActivities(
+            connection,
+            record.datasetId(),
+            record.simulationStartTime());
+        return activities.getLeft();
+      } catch (SQLException e) {
+        throw new RuntimeException(e);
+      }
+    }
+
+    @Override
+    public Instant startTime() {
+      return record.simulationStartTime().toInstant();
+    }
+
+    @Override
+    public Duration duration() {
+      return Duration.of(
+          record.simulationStartTime().microsUntil(record.simulationEndTime()),
+          Duration.MICROSECONDS);
     }
   }
 }
